+++ src/Line/EmptyLine.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilCompilationSource;
+namespace webignition\BasilCompilationSource\Line;
 
 class EmptyLine extends AbstractLine implements LineInterface
 {
