+++ tests/Unit/LineListTest.php
@@ -4,16 +4,16 @@ declare(strict_types=1);
 
 namespace webignition\BasilCompilationSource\Tests\Unit;
 
-use webignition\BasilCompilationSource\ClassDependency;
+use webignition\BasilCompilationSource\Line\ClassDependency;
 use webignition\BasilCompilationSource\ClassDependencyCollection;
-use webignition\BasilCompilationSource\Comment;
-use webignition\BasilCompilationSource\EmptyLine;
+use webignition\BasilCompilationSource\Line\Comment;
+use webignition\BasilCompilationSource\Line\EmptyLine;
 use webignition\BasilCompilationSource\LineListInterface;
 use webignition\BasilCompilationSource\Metadata\MetadataInterface;
-use webignition\BasilCompilationSource\Statement;
+use webignition\BasilCompilationSource\Line\Statement;
 use webignition\BasilCompilationSource\LineList;
 use webignition\BasilCompilationSource\Metadata\Metadata;
-use webignition\BasilCompilationSource\StatementInterface;
+use webignition\BasilCompilationSource\Line\StatementInterface;
 use webignition\BasilCompilationSource\VariablePlaceholderCollection;
 
 class LineListTest extends \PHPUnit\Framework\TestCase
